move token to impl package
