+++ sharding-core/sharding-core-rewrite/src/test/java/org/apache/shardingsphere/core/rewrite/token/pojo/impl/EncryptLiteralAssignmentTokenTest.java
@@ -15,33 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rewrite.token.pojo;
+package org.apache.shardingsphere.core.rewrite.token.pojo.impl;
 
-import lombok.Getter;
+import org.junit.Test;
 
-/**
- * Aggregation distinct token.
- *
- * @author panjuan
- */
-public final class AggregationDistinctToken extends SQLToken implements Substitutable {
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+
+public final class EncryptLiteralAssignmentTokenTest {
     
-    private final String columnName;
-    
-    private final String derivedAlias;
-    
-    @Getter
-    private final int stopIndex;
-    
-    public AggregationDistinctToken(final int startIndex, final int stopIndex, final String columnName, final String derivedAlias) {
-        super(startIndex);
-        this.columnName = columnName;
-        this.derivedAlias = derivedAlias;
-        this.stopIndex = stopIndex;
-    }
-    
-    @Override
-    public String toString() {
-        return null == derivedAlias ? columnName : columnName + " AS " + derivedAlias;
+    @Test
+    public void assertToString() {
+        EncryptLiteralAssignmentToken actual = new EncryptLiteralAssignmentToken(0, 1);
+        actual.addAssignment("c1", "c1");
+        actual.addAssignment("c2", 1);
+        assertThat(actual.toString(), is("c1 = 'c1', c2 = 1"));
     }
 }
